+++ CHANGELOG.md
@@ -15,6 +15,7 @@ guarantees will be maintained within each major version series.
 successful completion (i.e. normal BamWriter destruction, not triggered by a
 thrown exception) the file is renamed to the actual requested filename.
 - PBI file creation follows the same temporary naming convention.
+- Support for barcode pair (forward, reverse) in DataSetXML filter.
 
 ### Fixed
 - Improper 'clip to reference' product for BamRecord in some cases.